Make messages inherently persistent
The idea of a "persistent queue" yields poor separation of concerns.
If a message, upon creation, is automatically persisted to the
filesystem, the only thing the queue needs to do is delete the
message when it gets acknowledgement from the kafka producer.

This required all JSON marshaling to be done up front rather than at
the end of the pipeline. Hence it affected package schema. Messages
are concrete things now, not interfaces.

The persistence layer is currently in package kafka, which may or may
not be the best place for it. In any event, now that the persistor
exists, it could easily be moved elsewhere.