Add remote logger
The remote logger opens a socket and parses the incoming stream into
newline-delimited messages. Each message is converted to a Kafka
message by adding appropriate metadata and being marshaled to JSON.
This creates a message stream parallel to that generated by
instrumentation. The two streams are merged together before the queue.