agent: use anonymous sockets
The client has used named sockets to communicate with the child process. This
presents a security vulnerability, which is that an untrusted process could
connect to the socket before the child process. As only one connection is
served, the untrusted process would prevent the child from communicating, in
addition to gaining unauthorized access to Auklet's pipeline.

Anonymous sockets, which have no address, eliminate this vulnerability.