message: implement protobuf adapter
message/pbadapter.go

We implement a stage message.PBAdapter, that translates messages from
JSON to protobuf encoding. This requires that messages encode the
original Go type of the message; otherwise, the translator has no way
to choose the correct type to decode into.

The adapter is not used because there is no MQTT producer yet.

broker/message.go

This requires that broker.Message encode the original Go type prior
to encoding. The field "Type", which was previously used to associate
the message with a broker topic, now stores the original Go type. A
new field "Topic" stores the topic.

schema/log.go

As all messages now have to be defined in terms of protobuf
declarations, we introduce a type schema.AgentLog, which is currently
not consumed and thus has no specified schema, but which existed
previously, and continues to be transmitted. To support the protobuf
adapter, we have to introduce this type; otherwise these messages
simply cannot be encoded as protobuf.

schema/schema.proto

To clarify our intent, message declarations that are only used within
another message, and not as the top-level schema, are nested within
their "parents".