message: let PBAdapter satisfy only MessageSource
The previous idea was to translate messages from JSON to protobuf late
in the pipeline, after the queue. This required that the PBAdapter
satisfy the same interface as the queue, as expected by producers.

However, translation affects network bandwidth usage, and thus needs
to occur prior to data accounting. This means PBAdapter only has to
implement MessageSource, not the queue's full interface.