broker: move retry logic into producer
Previously, the queue and producer were jointly involved in retrying
messages that could not be sent. The queue would continue sending a
message until it got a dequeue request from the producer, which
indicated successful transmission.

This required a special interface between queue and producer that
compromised the pipeline's composability. Moving retry logic into the
producer means that the queue no longer needs to listen for dequeue
requests. Now, stages can be rearranged without modification.

If we use buffered channels, the queue has very little reason to
exist. The benefit of having an infintely expandable buffer seems
negligble; a buffer size of 10 has been give to the merger. To
compensate for the finite buffer size, the producer will only retry a
finite number of times before trying another message. The message is
not dropped; it will be retried on next startup. This retry logic
prevents the pipeline from being clogged indefinitely.