agent: refactor Server and Logger
Moving socket pair creation into package app improves the testability
of package agent. This allows better separation of concerns within
package agent as well; the code for requesting emissions from the
agent is now separate from the code that receives the emissions.
Both agent.Server and agent.Logger take in io.Reader types, improving
testability.

A small, somewhat unrelated change is to launch goroutines within
constructors (within package message), so that they do not have to be
explicitly launched in main. This improves readability and gives up
some flexibility that was not necessary.