agent: move handlers into package schema
Package agent no longer requires that any handler functions be
provided to it. Instead, it sends its own message format that can be
processed by another pipeline stage. Package schema implements a stage
that does this. This means the handlers no longer must be defined in
main.

The goal of this change is to make package agent easier to test.